oracle: add three-point (SSR/hydrated/csr-mount) and hydration-snap oracles
fixture-hydrate.playwright.ts only ever checks one thing: does THIS fixture's
frozen HTML+client-JS pair hydrate and react correctly. It never asks whether
the same fixture rendered a different way (deferred hydration, or a real
createComponent() CSR mount with no SSR input at all) produces the same
result — exactly the class of bug a hydration mismatch or a CSR-only code
path introduces silently, and exactly what the SSR-vs-CSR-mock divergence in
csr-render.ts can't catch since it never touches a real browser DOM.

normalizeHTML/stripConditionalMarkersForCrossAdapter move out of
jsx-runner.ts into html-normalize.ts: jsx-runner.ts imports bun:test at
module scope, which Playwright's Node worker_threads can't resolve even
though the top-level `bunx playwright` process runs under Bun. Existing
bun-test consumers keep importing from jsx-runner.ts unchanged (re-exported).

A first run against the 37-fixture corpus surfaces 18 real SSR/hydration
divergences the fixture-hydrate layer never checked for (reactive
boolean-attribute resets, rest-spread attribute loss, portal-origin-marker
timing, and an unexplained .value expando on non-form roots) — fixing them
is out of scope here, so each is quarantined in oracle-quarantine.ts with a
rot check that turns red the moment it starts passing again, the same
discipline csr-skip-set.ts/csr-skip-rot.test.ts already use for CSR
conformance.
